+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project aims to adhere to [Semantic Versioning](http://semver.org/).
 
+### Added 
+ - [Prune Empty Parent Directory](https://github.com/joyent/java-manta/issues/425)
+ 	Can now attempt to delete empty parent directories within a delete. See 
+ 	`manta.prune_empty_parent_depth`/`MANTA_PRUNE_EMPTY_PARENT_DEPTH` in the configuration documentation
+ 	for an explanation of how this can be enabled.
+
 ## [3.2.3] - 2018-08-03
 ### Fixed
  - [UnsupportedOperationException when getting a 0 byte file using `MantaClient.getAsInputStream`](https://github.com/joyent/java-manta/issues/408)
@@ -426,3 +432,7 @@ Version skipped due to a release error.
 ### Fixed
  - Fixed checkstyle failures.
  - Fixed pom.xml settings so that it will properly release to Maven Central.
+ 
+ 
+
+